+++ src/main/java/domain/Lotto/LottoNumberFactory.java
@@ -1,10 +1,10 @@
-package domain;
+package domain.Lotto;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class LottoFactory {
+public class LottoNumberFactory {
 
     private static final int MINIMUM_LOTTO_NUMBER = 1;
     private static final int MAXIMUM_LOTTO_NUMBER = 45;
