+++ src/main/java/domain/LottoGenerator/WinningLottoGenerator.java
@@ -1,22 +1,18 @@
-package domain;
+package domain.LottoGenerator;
 
-import java.util.Collections;
-import java.util.Comparator;
+import domain.Lotto.Lotto;
+import domain.Lotto.LottoNumber;
+import domain.LottoGenerator.LottoGenerator;
+
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class AutoLottoGenerator implements LottoGenerator {
-
-    private static final int LOTTO_SIZE = 6;
+public class WinningLottoGenerator implements LottoGenerator {
 
     @Override
     public Lotto generateLotto(List<Integer> numbers) {
-        Collections.shuffle(numbers);
-
         List<LottoNumber> lottoNumbers = numbers.stream()
-                .limit(LOTTO_SIZE)
                 .map(LottoNumber::new)
-                .sorted(Comparator.comparing(LottoNumber::number))
                 .collect(Collectors.toList());
 
         return new Lotto(lottoNumbers);
