+++ src/main/java/domain/player/Money.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 import utils.ExceptionMessage;
