+++ src/test/java/domain/LottoNumberTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.Lotto.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
