+++ src/test/java/domain/MoneyTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.player.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import utils.ExceptionMessage;
